requirements.txt
matplotlib~=3.5.2
numpy~=1.22.1
pandas~=1.2.4
pydicom==2.3.0
opencv-python~=4.5.5.64
scikit-image~=0.18.1
Pillow~=9.0.1
tifffile~=2021.4.8